Seperate Engine from Github Files
The following change seperates the important aspect of the engine and its files from the files of github such as readme, git ignore etc.  Ignore can remain in the root considering it could potentially check for files greated in the directory root not specifically the engine root itself,, where it gives a possibility for changes.

:+1: