Change Build Path to Solution Dir From Proj Dir
The following addressed the previously mentioned by @JDSherbert , that the build should be in the Solution directory rather than with the project files.

Which is no addressed
the issue was with the mistake done in properties-> Configuration,.,. -> General -> Intermediate Directory = $(SolutionDir)$(Platform)\$(Configuration)\  , which was not defined with SolutionDir, by default it took  " projdir/x64/debug etc.