Build Events for Debug / Release (DLLs)
Additionally added a folder with specified DLLs for the platforms of Debug and Release.

Now both are checked depending on the platform and the required DLLs are additionally added to the build folders depending on which version is currently being built.

Please Note the specified area for PreLink Events:

ping -n 6 127.0.0.1

The following information can be read about here :
https://github.com/VerzatileDevOrg/VS.DLLinFolder